Updated Examples to work with armv7+armv7s

+++ Examples/cocos2d-iphone-1.0.1/cocos2d/CCActionInstant.m
@@ -257,7 +257,7 @@
 {
 	return [NSString stringWithFormat:@"<%@ = %08X | Tag = %i | target = %@ | selector = %@>",
 			[self class],
-			self,
+			(NSUInteger)self,
 			tag_,
 			[targetCallback_ class],
 			NSStringFromSelector(selector_)
