+++ Examples/cocos2d-iphone-1.0.1/cocos2d/CCTextureAtlas.m
@@ -110,7 +110,7 @@
 
 - (NSString*) description
 {
-	return [NSString stringWithFormat:@"<%@ = %08X | totalQuads =  %i>", [self class], self, totalQuads_];
+	return [NSString stringWithFormat:@"<%@ = %08X | totalQuads =  %i>", [self class], (NSUInteger)self, totalQuads_];
 }
 
 -(void) dealloc
